withRequestParams() makes @param 'queries' optional
Before this, withRequestParams() assumed it always received 'queries'
arg.
Decorators such as withSpinner() assumed `load()`'s signature did not
include queries.  So when it passed the cb in place of queries and so
cb would never be called.  This was causing the spinner to always
display
in ActiveIconViews other than Search.

Minor details:

* <Spinner />, get rid of sylizing used for debugging purposes
* withSpinner(), show spinner at all times when there's data loading